#4 Changed

+++ NativeStructs/MONITORINFO.cs
@@ -6,9 +6,9 @@ namespace PinkWpf.NativeStructs
     [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto, Pack = 4)]
     public class MONITORINFO
     {
-        public int cbSize = Marshal.SizeOf(typeof(MONITORINFO));
+        public uint cbSize = (uint)Marshal.SizeOf(typeof(MONITORINFO));
         public Win32Rect rcMonitor;
         public Win32Rect rcWork;
-        public int dwFlags;
+        public uint dwFlags;
     }
 }
